chore: update src

+++ helpers/extract_embeddings.py
@@ -10,29 +10,33 @@ from tqdm import tqdm
 import numpy as np
 from segment_anything import sam_model_registry, SamPredictor
 
+
 def main(checkpoint_path, model_type, device, images_folder, embeddings_folder):
     sam = sam_model_registry[model_type](checkpoint=checkpoint_path)
     sam.to(device=device)
     predictor = SamPredictor(sam)
 
     for image_name in tqdm(os.listdir(images_folder)):
+        out_path = os.path.join(embeddings_folder, os.path.splitext(image_name)[0] + ".npy")
+        if (not image_name.endswith(".png")) or (os.path.exists(out_path)):
+            continue
+
         image_path = os.path.join(images_folder, image_name)
         image = cv2.imread(image_path)
         image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-
 
         predictor.set_image(image)
 
         image_embedding = predictor.get_image_embedding().cpu().numpy()
 
-        out_path = os.path.join(embeddings_folder, os.path.splitext(image_name)[0] + ".npy")
         np.save(out_path, image_embedding)
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--checkpoint-path", type=str, default="./sam_vit_h_4b8939.pth")
-    parser.add_argument("--model_type", type=str, default="default")
-    parser.add_argument("--device", type=str, default="cuda")
+    parser.add_argument("--model_type", type=str, default="vit_h")
+    parser.add_argument("--device", type=str, default="cpu")
     parser.add_argument("--dataset-path", type=str, default="./example_dataset")
     args = parser.parse_args()
